# IQ3_S improvements