This PR makes some improvements to the `IQ3_S` quants. Nothing earth shuttering, but I think still worth having. We get
* Lower quantization error
* Some speedup on all platforms
* About 40% decrease in quantization time

This table shows a comparison between the quantization error on master and this PR. The quantization error is defined as `PPL(IQ3_S)/PPL(fp16) - 1`. The context is 4096 for LLaMA-v2 and 2048 for LLaMA-v1

| Model |  Master |  PR |  Reduction |
|---:|---:|---:|---:|
|LLaMA-v1-7B   | 4.53% | 4.01% | 0.885 |
|LLaMA-v2-7B   | 4.66% | 4.03% | 0.865 |
|Mistral-7B         | 3.08% | 3.13% | 1.016 |
|LLaMA-v1-13B | 3.34% | 2.99% | 0.895 |
|LLaMA-v2-13B | 3.34% | 3.00% | 0.898 |

This table shows a comparison in inference speed on the platforms where I can test:
* CUDA on RTX-4080
* AVX2 on Ryzen 7950X
* ARM_NEON on M2 Max
* Metal on 30-core M2 Max GPU

| model          | backend    |    threads | test       |     t/s (Master) |         t/s (PR) |   Speedup |
| ---------------| ---------- | ---------: | ---------- | ---------------: | ---------------: | --------: |
| llama 7B IQ3_S | CUDA       |          1 | pp 512     |    5699.0 ± 16.9 |    5722.9 ± 18.9 |  1.004    |   
| llama 7B IQ3_S | CUDA       |          1 | tg 128     |    151.53 ± 0.13 |    152.68 ± 0.16 |  1.008    |   
| llama 7B IQ3_S | AVX2       |         16 | pp 512     |     23.38 ± 0.35 |     28.55 ± 0.15 |  1.221    |   
| llama 7B IQ3_S | AVX2       |         16 | tg 128     |     15.98 ± 0.27 |     16.19 ± 0.22 |  1.013    |   
| llama 7B IQ3_S | AVX2       |          8 | tg 128     |     11.71 ± 0.05 |     13.86 ± 0.03 |  1.184    |   
| llama 7B IQ3_S | Metal      |          4 | pp 512     |    457.99 ± 0.29 |    457.75 ± 0.26 |  1.000    |   
| llama 7B IQ3_S | Metal      |          4 | tg 128     |     49.69 ± 0.01 |     50.35 ± 0.01 |  1.013    |   
| llama 7B IQ3_S | ARM_NEON   |          8 | tg 128     |     10.01 ± 0.01 |     11.00 ± 0.03 |  1.100    |   
| llama 7B IQ3_S | ARM_NEON   |          4 | tg 128     |      5.18 ± 0.00 |      5.73 ± 0.00 |  1.106    | 

ARM performance is still pathetic. Metal performance is acceptable (0.95X for TG and 1.02X for PP compared to `Q3_K_S`). AVX2 Performance is acceptable for TG (0.88X compared to `Q3_K_S`) but bad for PP (0.38X of `Q3_K_S`. I guess, for PP the loading of the lattice points from the lookup table becomes I/O bound and severely limits performance). CUDA performance is excellent (1.04X/1.015X for TG/PP compared to `Q3_K_S`, 1.18X compared to `Q4_0` TG). 

The main reason this PR exists is @PeterReid's suggestion to replace the lookup table with a multiplier that generates the codebook. While working on that I noticed that there is still room for improvement for the lookup based `IQ3_S`, and this PR is the result. I did find a few multipliers that work quite OK (quantization error about 1-1.5% higher than lookup), but there was no performance improvement on CUDA and TG on AVX2, so I decided to not further pursue this idea.