@Nindaleth 2024-03-02
> I decided to not further pursue this idea

How did the actually slow archs do with the other approach? I be misunderstanding the description, but it sounds to me as AVX2 PP and ARM PP/TG performance is bad with the lookup approach, but you decided to not look into the multipliers any further because the already good CUDA PP/TG and AVX2 TG performance with lookup didn't get any better with multipliers...?