@Artefact2 2024-03-02
I notice you change `dequantize_row_iq3_s`, are older IQ3 GGUFs still useable or is this breaking backward compatibility?