@ikawrakow 2024-03-02
> I notice you change the grid in `dequantize_row_iq3_s`, are older IQ3 GGUFs still useable or is this breaking backward compatibility?

Sorry, forgot to mention. Yes, previous quantization is still usable. You can re0quantize to take advantage of the slight PPL improvement.