@Nindaleth 2024-03-02
> because the quantization error about 1-1.5% higher than lookup

As in one percentage point higher, not 1.01x higher? I see, that makes sense, thanks for pointing that out.