@ikawrakow 2024-03-02
If the quantization error is 3% with lookup, it will be in the 4 to 4.5% range with the best multiplier I was able to find. Example: PPL = 5 for `fp16`, PPL = 5.15 with lookup, PPL = 5.20-5.23 with multiplier.