@JianbangZ 2024-03-12
@ikawrakow I was testing some iq3_m quant today I found the PPL was not as good as before. After digging into your recent PRs, I think this grid change to iq3_s causes a PPL increase for my model as iq3_m uses a log of iq3_s. (i use qwen1.5-7b-chat for testing)

So seems this grid value is pretty important to the final perforhamce for IQ quantization method, and I am seeing opposite results compared to yours. 

Would you mind elobrating how you choose gird values?