@ikawrakow 2024-03-12
@JianbangZ Sorry to hear that this PR is increasing PPL for your model. It did improve PPL on 8 of the 9 models tested, and it was about the same for one.

Concerning grid points: during quantization the quants can take one in 8 values in `{1, 3, 5, 7, 9, 11, 13, 15}`. This has not changed (and neither has the codebook that selects 512 out of the 4096 possible values for a group of 4 quants). During inference, prior to this PR the values used were `{4, 12, 20, 28, 36, 44, 52, 62}` with an additional factor of 1/4 applied, so effectively `{1, 3, 5, 7, 9, 11, 13, 15.5}`. These are encoded in `iq3xs_grid` (which became `iq3s_grid` in this PR), where each `uint32_t` entry contains 4 quants. I had found empirically that using a slightly larger value for the highest quantization value  improves PPL. This PR changed the quant values back to `{1, 3, 5, 7, 9, 11, 13, 15}`, but added a fudge factor of `1.033` to the super-block scale. This change did improve PPL on 8 out of 9 tested models. The second change that may affect your result is a reduction in the scale search range during quantization. Before this PR, the scale search range was `-15...15`, and became `-9...9` in this PR. This did not have any effect on the PPL of the 9 tested models, but reduced quantization time by about ~40%. The decision to make this change was based on the feedback from many that `IQ` quantization takes too a long time. Checking if this search range reduction has negatively impacted your model is very easy to do: just go to [this line](https://github.com/ggerganov/llama.cpp/blob/306d34be7ad19e768975409fc80791a274ea0230/ggml-quants.c#L11257) in the source code, change to `for (int is = -15; is <= 15; ++is) {`, recompile, re-quantize, and see what happens.