@JianbangZ 2024-03-12
> > I notice you change the grid in `dequantize_row_iq3_s`, are older IQ3 GGUFs still useable or is this breaking backward compatibility?
> 
> Sorry, forgot to mention. Yes, previous quantization is still usable. You can re0quantize to take advantage of the slight PPL improvement.

Since grid changed, and dequant seems use the new grid value, how was the the old grid quant model compatiable?