@Nexesenex 2024-04-05
@ikawrakow Could you make a different GGML_type of the old IQ3_S, so we can quantize with it?
I excluded this one from my quant cocktails because it's behaving very poorly on Miqu 70b : the IQ2_S and IQ2_M at least are bungled by its presence by a 10% perplexity jump compared to expectations, and I never observed any PPL gain with it anyway. When I change the tensors which are in IQ3_S toward IQ3_XXS, I revert to expected results in line with IQ2_XS.

Or maybe this PR https://github.com/ggerganov/llama.cpp/pull/5834 is causing problems (I don't know if I tested this new IQ3_S with ot without the PR back in the days) ?