@Nexesenex 2024-04-05
@ikawrakow Could you make a different GGML_type of the old IQ3_S, so we can quantize with it?
I excluded this one from my quant cocktails because it's behaving very poorly on Miqu 70b : the IQ2_S and IQ2_M at least are bungled by its presence by a 6-7% perplexity jump compared to expectations, and I never observed any PPL gain with it anyway. When I change the tensors which are in IQ3_S toward IQ3_XXS (and Q2_K for the embedding.weight), I revert to expected results at least a bit better than with IQ2_XS.

At PPL 512 :

Miqu IQ2_M : approx 4.62 (stopped around 500 chunks)
Miqu IQ2_M_fixed : 4.3371
Miqu IQ2_S : 4.7721
MIqu IQ2_S_fixed : 4.4912
Miqu IQ2_XS for information : 4.5553

Or maybe this PR https://github.com/ggerganov/llama.cpp/pull/5834 is causing problems (I don't know if I tested this new IQ3_S with it without the PR back in the days) ?